requirements.txt
# V2R 프로젝트 Python 의존성

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# 네트워크 스캐닝
python-nmap>=0.7.1
requests>=2.31.0

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0

# 데이터베이스
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0

# ML/LLM
openai>=1.0.0
scikit-learn>=1.3.0
# xgboost는 CPU 전용으로 설치 (GPU 의존성 제거)
xgboost>=2.0.0

# 리포트 생성
python-docx>=1.1.0
reportlab>=4.0.0

# 웹 대시보드
streamlit>=1.28.0
flask>=3.0.0

# 유틸리티
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
docker>=6.1.0

# 테스트
pytest>=7.4.0
pytest-cov>=4.1.0

